Add a profile preference to track the library selection screen
This adds a flag to the profile preferences that is used to track
whether or not the user has seen the library selection screen that's
typically opened on application startup. It also adds a switch to
the developer settings menu to toggle the flag for testing purposes.

Affects: https://jira.nypl.org/browse/SIMPLY-2718