Remove "the current account"
This makes the changes necessary to remove the concept of "the
current account" from the application. From now on, no account in the
application is considered to be "the current account"; selecting the
catalog tab either shows you the account you viewed most recently, or
arbitrarily selects the first (in account ID order) account from the
current profile. Feed have been augmented with ownership information: A
feed either belongs to a given account (and this ownership information
is used to look up credentials when borrowing books or loading the
next parts of paginated feeds), or each entry in the feed specifies
the account to which the entry belongs. This eliminates many bugs
related to the fragment backstack - it's possible that "the current
account" may have changed upon navigating back to a feed fragment on
the backstack, and so subsequent account operations performed by that
fragment would have been executed using the wrong account.

This also removes the NYPL-compatible behaviour of only showing books
and holds from "the current account", instead providing facets to
filter books shown in the Books and Holds tabs.

Affects: https://jira.nypl.org/browse/SIMPLY-2767