Refactor the account screen (again)
This adjusts the UI changes made in the previous commit to use a single
login button (instead of a possibly-image-button view pair), and adds
a new "waiting for external authentication" login state. The intention
is that the OAuth login process will call profileAccountLogin() as
normal, and that method will set the account login state to "waiting
for external authentication" and return immediately (as if the entire
login process had finished). The settings fragment can then either
complete the process upon receipt of an OAuth token, or cancel the
login process entirely.

Affects: https://jira.nypl.org/browse/SIMPLY-2224