Use the browser for OAuth
This publishes an Intent that takes the user through the browser-based
OAuth workflow. It parses the Intent returned by the browser and gives
the resulting authentication token to the profile controller in order
to set the account as logged-in. Currently, the profile controller code
just raises an UnimplementedCodeException, as finishing the controller
code will involve refactoring the AccountAuthenticationCredentials
type to cope with the fact that not every authenticated user has a
username and password.

This requires applications to specify an intent-filter with a custom
URI scheme in their manifests, and to publish that same schema via
the BuildConfigurationServiceType class. Note that applications _must_
now publish a build configuration service; the core module no longer
provides a fallback module as there are no sensible defaults it can
provide here.

Affects: https://jira.nypl.org/browse/SIMPLY-2794
Affects: https://jira.nypl.org/browse/SIMPLY-2795