[SIMPLY-2773] Fixes an issue where the default account did not match what the patron expected.
When first adding an account, the patron could be shown the SimplyE
catalog, instead of the library they just added. This would result in
the age gate prompt being presented to the patron, seemingly in error.

This fixes the issue by presenting the first non-default account when we
lack a most recent account id.