Make this code more explicit.
I accidentally removed this code because the value was unused and I did
not realize it had necessary side effects. This broke the app and it
took me a while to realize what I'd done.

To make this more explicit I've rewritten the comment and removed the
unused account value.