Hide R2 behind a profile preference
This adds a new profile preference that can be toggled via the
developer menu. If the preference `useExperimentalR2` is set to
`true`, then DRM-free books will be opened in R2. Otherwise, both
DRM-free and DRM-encumbered books will be opened in R1. When
R2 stops being experimental, all of this code can be removed.