Fix a possible crash when the main fragment is detached from the activity.
When starting the main fragment We post a delayed runnable to create the
tabbed navigation controller. It's possible that this could execute
after the app has been backgrounded and the fragment detached. This
could cause a crash, as seen in Firebase.

Mitigate this crash by checking for a non-null activity context before
executing the runnable.