Hide the up affordance for the ReaderActivity.
The reader has no canonical 'parent' as it could feasibly be started by
many different parents. Thus, according to the Android guidelines the up
navigation should not be shown. Instead, the system back should be used
to exit the reader.

If this is not acceptable and the up affordance must be shown we should
use an 'X' icon to indicate you're closing the reader instead of the
arrow icon.