SIMPLY-2847: Fix overlay issue with toolbar and bookmarks fragment.
Added a separate fragment container with 'fitsSystemWindows=true'. This
causes the bookmarks fragment to show properly with the immersive window
flags set on the reader activity.

Additionally, this change hides the reader fragment, instead of
replacing it. This means the reader views are not removed from the
container and don't need to be added again by the system when popping
the back stack.