Continue task recorder refactoring
This eliminates the "error" type parameter from the task recorder and
refactors all existing code to use the new lesser-typed interface.
It adjusts the task recorder API such that Presentable attributes
are no longer an aspect of individual task steps but of the recorder
itself (and, as a result, of the task result). It also simplifies the
Presentable interfaces and removes functionality that we weren't using.

This also deletes the contents of the sandbox module.

Affects: https://jira.nypl.org/browse/SIMPLY-3001
Affects: https://jira.nypl.org/browse/SIMPLY-2940