Remove the tests-android and tests-strings modules
Previously the test suite was structured such that each test class
was actually three classes:

  1. An abstract "Contract" class in simplified-tests/src/main/java
     containing the actual tests.
  2. A concrete "Test" class in simplified-tests/src/test/java
     that runs the Contract on the local machine without a real
     or emulated Android device.
  3. A concrete "Test" class in simplified-tests-android/src/test/java
     that runs the Contract on a real or emulated Android device.

This was necessary because of the brain-dead way that Android decided
to implement testing: In order for a test to run on a device, the test
must use a proprietary instrumentation API. This means that a test
class must _either_ be a local unit test, _or_ a device unit test,
but cannot be both. The Contract/Test split was created to work around
this stupidity.

The problem; maintaining device tests is a full-time job. The test
runner for Android is fragile to the point of being essentially
useless. We didn't maintain the tests, and nobody had run them for
quite some time.

This change simply removes the tests-android module. The tests are
still split into Contract/Test classes, but these now all live in
simplified-tests/src/main/java and we only support running tests
locally. New unit tests do not need to have a Contract/Test split.