Fix Adobe deactivations
This appears to fix the long-standing issue of being unable to deactivate
Adobe DRM devices on Android. It turned out that the code was correct, but
that it was necessary to perform the device deactivation using a fresh
client token. This was discovered after @ettore noticed some odd calls
the iOS client was making, and this lead to a discussion with @leonardr
in the ticket https://jira.nypl.org/browse/SIMPLY-3208 that lead me to
try deactivation with a fresh token. I've so far been able to deactivate
the device every time, with no sign of the dreaded E_ADEPT_LOGIN_ERROR.

Thanks @ettore, @leonardr!

Affects: https://jira.nypl.org/browse/SIMPLY-3208
Affects: https://jira.nypl.org/browse/SIMPLY-3099
Affects: https://jira.nypl.org/browse/SIMPLY-2980