Make ServiceDirectoryProviderType.serviceDirectory() block
This switches to a model where asking for the service directory is now
a blocking operation. In theory, we block indefinitely.  In practice,
we place a three minute bound on the operation: If it's taking longer
than this to retrieve the service directory, something is seriously
wrong and we should abort, print an stacktrace, and then throw an
exception.

This should fix any and all problems where we get a lifecycle-related
exception, and may also remove the need to use Kotlin `lazy` properties.