Delete simplified-http
This deletes the simplified-http and removes all remaining references
to it in the codebase. This required replacing the contents of
the simplified-documents module with a somewhat more rudimentary
implementation. We also deleted the simplified-clock module. It also
required introducing an intermediate factory type for the accounts
source SPI, because the NYPL registry needs access to the HTTP client
and this can't be instantiated from ServiceLoader.

Affects: https://jira.nypl.org/browse/SIMPLY-3140