Remove boot string resources
This removes a ton of boot string resources that we'll never need
to translate. This reduces the translation burden for projects like
LFA, and removes a lot of boilerplate code.

It also took the opportunity to clean up the boot process, which
had a couple of instances of submitting services multiple times.