Attempt to keep annotations URIs
The account provider resolution code wasn't taking advantage of the
fact that an annotations URI _might_ be included in an account provider
description if that description was generated from an existing account
provider. This fixes that, so we should be keeping hold of annotations
URIs for as long as we can!

Affects: https://jira.nypl.org/browse/SIMPLY-2403