Fix a crash for Related Books...
This fixes a crash triggered when the user follows a Related Books
link from a book detail page when that book detail page came from
a book in "My Books". This fix introduces a new `resolveFeedFromBook`
method that requires passing in explicit ownership information for
the feed. This is somewhat of a stopgap measure. The real fix is
to replace the 1:1 correspondence between catalog feed fragments
and catalog view models with an N:1 correspondence; catalog feed
and book detail fragments should share a single view model that
stores the entire stack of feeds and book detail entries. Once
that happens, full ownership information will be available in
all circumstances and methods like `resolveFeedFromBook` will
no longer be necessary.

Fix: https://jira.nypl.org/browse/SIMPLY-3471