Don't redirect to settings for no-auth accounts
The current application behaves badly if the server hosting an OPDS
feed returns a response that indicates that authentication is required,
but the corresponding authentication document for the account contains
no authentication mechanisms. What will happen currently is that the
app effectively goes into an unbreakable loop of redirecting the user
to the settings screen and then not presenting any way for the user
to actually log in. This change makes it so that if the account has
no authentication mechanisms, we just show a feed error instead of
opening the settings screen.