Explicitly return to catalog instead of popping backstack
This adjusts the behaviour of the account fragment in that the "closeOnLoginSuccess"
flag now explicitly switches back to the catalog tab. This is a bit of a stopgap
measure to specifically support Open eBooks and should be replaced when we have
a better viewmodel backing the catalog (the viewmodel would subscribe to account
events, and would refresh itself when an account logs in).

Fix: https://jira.nypl.org/browse/SIMPLY-3476