Implement the full Simplified-Bookmarks-Spec
This implements the remainder of the Simplified Bookmarks spec [0],
including unit tests to demonstrate that the expected test samples
are parsed the correct way.

The bookmarks spec has been imported as a submodule so that the
various included test cases can be used as test suite resources
directly.

We now enforce that all bookmarks must be specified in a UTC
time zone (although they can be displayed to the user in any
time zone).

[0] https://github.com/NYPL-Simplified/Simplified-Bookmarks-Spec

Affects: https://jira.nypl.org/browse/SIMPLY-3605