Introduce a noise-reduction policy
This moves logging configuration into its own class, and reduces the
logging levels of some of the noisier parts of the app. This appears
to improve the UI performance of the app somewhat, but I'm willing
to accept that it might be a placebo effect.

Affects: https://jira.nypl.org/browse/SIMPLY-3584