Introduce annotations URI as account credential
This adds a field to account credentials that stores the annotations
URI. This URI is taken from the patron user profile during the login
process. The URI is currently not used anywhere in the application -
the old (now deprecated) URI is used instead.

Affects: https://jira.nypl.org/browse/SMA-35