Fix status in the registry after revocation by modifying BookRegistry's concerns.
Books are not removed any more from the registry after deletion so that opened fragments can carry on keeping track of book status.
FeedLoader doesn't override loaded feeds with BookRegistry data any more. On the contrary, it updates the status of books already available in the registry with fresh data.