Introduce UIExecutor to fix a bug occuring when removing an account
Deleting an account could cause a crash in case there were two library accounts: the SimplyE collection and another one.
When deleting the latter account while it was shown in the Catalog tab, the CatalogFeedFragment associated with it happend to receive the callback
onAgeUpdated meant to the CatalogFeedFragment showing the SimplyE collection before being deleted. Then, as the account had been deleted, it couldn't be found any more and the app crashed.