Produce a QA Open eBooks build
This switches the Open eBooks app (temporarily) to using the QA
server for testing. This also changes the colour scheme to a pink
colour so that it's clear immediately upon looking that this IS NOT
a production version for the store. Both the server and colour scheme
must be restored to normal before a release can be made.

Affects: https://jira.nypl.org/browse/SIMPLY-2214