Make the account bookmark sync switch reflect reality
This adjusts the account bookmark sync switch so that it works based
off of an observable status value. When the switch is toggled, the
reader bookmark service tracks the progress of the operation and can
publish appropriate status values when asked. A small indefinite
progress indicator has been added to show that a long-running
operation is occurring when the switch is toggled, and the switch is
marked as disabled (non-interactable) when a sync setting change is
in progress. This prevents users from being able to queue up dozens
of switch changes by hammering the toggle repeatedly.

Affects: https://jira.nypl.org/browse/SMA-186