Schedule "revocation" for expired loans
This change hopefully fixes the issue where books are deleted from
the on-disk database, but the book registry is not updated properly
to indicate that loans are gone. This change simply removes the
database entry call, and schedules a revocation of the loan. The
revocation code _should_ be intelligent enough to correctly do any
and all cleanup work required.

Affects: https://jira.nypl.org/browse/SMA-205