Use 'alternate' link for loan syncing
This undoes the previous commit and switches to a model where the book
registry for a given book is updated with up-to-the-minute-accurate
information on the server, using the "alternate" permalink for
each book. This should result in accurate book states when removing
books due to expired loans.

Affects: https://jira.nypl.org/browse/SMA-205