Implement an OE QA toggle
This implements a toggle that will add and remove an Open eBooks
QA account in the debug menu. Hardcoding the account information,
unfortunately, appears to be the only way to implement this as
there isn't a registry present in OE that seems to know about the OE
QA server.

Fix: https://jira.nypl.org/browse/OE-140