Delete book data for books with unsupported DRM systems
This updates the book database to be aware of book and DRM formats. If
an attempt is made to open a book that has an unsupported type of
DRM, the book data and the DRM information is deleted and the book is
reopened. The intention is that this will cause the book to appear
as having been loaned but not downloaded, and the user will then
re-download the book using the correct DRM system.  Bookmarks and
other information are preserved.

Affects: https://jira.nypl.org/browse/OE-192